+++ tests/Feature/Api/SitesTest.php
@@ -158,6 +158,30 @@ class SitesApiTest extends TestCase
         $response->assertJsonMissingValidationErrors(['name']);
     }
 
+    /** @test */
+    public function can_create_matching_user_when_updating_site(): void
+    {
+        $site = Site::create(['name' => 'Hello World']);
+
+        $response = $this->authed()->putJson('/api/sites/' . $site->id, [
+            'create_user' => true,
+            'name' => 'Hello World',
+            'document_root' => '/var/www/hello-world',
+            'primary_domain' => 'example.com',
+            'source_repo' => 'https://github.com/dshoreman/servidor-test-site.git',
+            'type' => 'basic',
+        ]);
+
+        exec('sudo userdel hello-world && sudo rm -rf /home/hello-world');
+
+        $response->assertOk();
+        $response->assertJsonStructure(['name', 'system_user', 'type']);
+        $this->assertArraySubset([
+            'dir' => '/home/hello-world',
+            'name' => 'hello-world',
+        ], $response->json()['system_user']);
+    }
+
     /**
      * @test
      * @depends authed_user_can_list_sites