+++ tests/Feature/Api/System/Groups/UpdateGroupTest.php
@@ -55,6 +55,55 @@ class UpdateGroupTest extends TestCase
     }
 
     /** @test */
+    public function authed_user_can_update_group_gid(): void
+    {
+        $group = $this->authed()->postJson($this->endpoint, [
+            'name' => 'gidtestgroup',
+        ])->json();
+
+        $response = $this->authed()->putJson($this->endpoint($group['gid']), [
+            'name' => 'gidtestgroup',
+            'gid' => (int) $group['gid'] + 1,
+        ]);
+
+        $response->assertOk();
+        $response->assertJsonFragment(['gid' => (int) $group['gid'] + 1]);
+        $response->assertJsonStructure($this->expectedKeys);
+
+        $this->addDeletable('group', $response);
+    }
+
+    /** @test */
+    public function authed_user_can_remove_all_users_from_group(): void
+    {
+        $group = $this->authed()->postJson($this->endpoint, [
+            'name' => 'groupwithmembers',
+        ]);
+
+        $this->addDeletable('group', $group);
+        $users = ['bin', 'daemon', 'games'];
+        $gid = $group->json()['gid'];
+
+        $group = $this->authed()->putJson($this->endpoint($gid), [
+            'name' => 'groupwithmembers',
+            'users' => $users,
+        ]);
+
+        $group->assertOk();
+        $group->assertJsonFragment(['users' => $users]);
+        $group->assertJsonStructure($this->expectedKeys);
+
+        $response = $this->authed()->putJson($this->endpoint($gid), [
+            'name' => 'groupwithmembers',
+            'users' => [],
+        ]);
+
+        $response->assertOk();
+        $response->assertJsonFragment(['users' => []]);
+        $response->assertJsonStructure($this->expectedKeys);
+    }
+
+    /** @test */
     public function cannot_update_nonexistant_group(): void
     {
         $response = $this->authed()->putJson($this->endpoint(9032), [
@@ -62,5 +111,61 @@ class UpdateGroupTest extends TestCase
         ]);
 
         $response->assertStatus(Response::HTTP_UNPROCESSABLE_ENTITY);
+    }
+
+    /** @test */
+    public function cannot_add_nonexistent_users_to_a_group(): void
+    {
+        $group = $this->authed()->postJson($this->endpoint, [
+            'name' => 'godfather',
+        ]);
+
+        $this->addDeletable('group', $group);
+
+        $response = $this->authed()->putJson($this->endpoint($group->json()['gid']), [
+            'name' => 'godfather',
+            'users' => ['ghost'],
+        ]);
+
+        $response->assertStatus(Response::HTTP_UNPROCESSABLE_ENTITY);
+        $response->assertJsonValidationErrors(['gid']);
+        $response->assertJsonFragment(['gid' => ["Couldn't update the group's users. Exit code: 3."]]);
+    }
+
+    /** @test */
+    public function updating_a_group_should_fail_if_name_exists(): void
+    {
+        $group = $this->authed()->postJson($this->endpoint, [
+            'name' => 'renamefail',
+        ]);
+
+        $this->addDeletable('group', $group);
+        $data = $group->json();
+        $data['gid'] = 3;
+
+        $response = $this->authed()->putJson($this->endpoint($group->json()['gid']), $data);
+
+        $response->assertStatus(Response::HTTP_UNPROCESSABLE_ENTITY);
+        $response->assertJsonValidationErrors(['gid']);
+        $response->assertJsonFragment(['gid' => ["Couldn't update the group. Exit code: 4."]]);
+    }
+
+    /** @test */
+    public function updating_a_group_without_changes_should_fail(): void
+    {
+        $group = $this->authed()->postJson($this->endpoint, [
+            'name' => 'changeless',
+        ]);
+
+        $response = $this->authed()->putJson(
+            $this->endpoint($group->json()['gid']),
+            $group->json(),
+        );
+
+        $response->assertStatus(Response::HTTP_UNPROCESSABLE_ENTITY);
+        $response->assertJsonValidationErrors(['gid']);
+        $response->assertJsonFragment(['gid' => ['Nothing to update!']]);
+
+        $this->addDeletable('group', $group);
     }
 }