+++ CHANGELOG.md
@@ -13,9 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 * Login page is now styled to match dark mode
+* Validation errors for system user's names are more specific again
 
 ### Fixed
+* All users can now be removed from a group, not just some
 * File browser now shows "???" when owner or group don't exist
+* Bulleted user lists under group names no longer update prematurely
 * Login form no longer breaks out of the layout to full page width
 * Stats bar no longer polls for data after logging out